refactor(jobs): small cleanups from PR review
- cli: promote default server URL literal to `DEFAULT_SERVER_URL` const
- store: extract `job_runs` DDL into `INIT_SCHEMA_SQL` const
- executor: swap `CancelFlag`'s `Mutex<bool>` for `AtomicBool` to drop
  the lock poisoning recovery and `*guard = true` dereferences
- destination: split the 403-line `destination.rs` into a submodule
  directory (`mod.rs` for trait + shared helpers, plus `lance.rs` and
  `sql_dml.rs` per backend) so the next destination impl (Iceberg,
  Delta, etc.) lands as its own file
